+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/kubernetes_kurtosis_backend.go
@@ -2,7 +2,6 @@ package kubernetes_kurtosis_backend
 
 import (
 	"context"
-	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/image_build_spec"
 	"io"
 
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend/engine_functions"
@@ -477,11 +476,6 @@ func (backend *KubernetesKurtosisBackend) CreateReverseProxy(ctx context.Context
 func (backend *KubernetesKurtosisBackend) DestroyReverseProxy(ctx context.Context) error {
 	// TODO IMPLEMENT
 	return stacktrace.NewError("Destroying the reverse proxy isn't yet implemented on Kubernetes")
-}
-
-func (backend *KubernetesKurtosisBackend) BuildImage(ctx context.Context, imageName string, imageBuildSpec *image_build_spec.ImageBuildSpec) error {
-	// TODO IMPLEMENT
-	return stacktrace.NewError("Building images isn't yet implemented in Kubernetes.")
 }
 
 // ====================================================================================================